requirements.txt
numpy~=1.26.4
PyQt5~=5.15.9
pillow~=10.3.0
tinydb~=4.8.0
Cython~=3.0.10
cx_Freeze~=7.0.0
PyQt-Fluent-Widgets
